Extract ResourceDiscount#matches? to ResourceMatchable module
This will be used in an upcoming ResourceCredit model.
